minor CSS adjustment in fullscreen media manager
This is an odd one. I had made this change already in 6adcbdbd and
(and dokufreaks/template-default@c7cf76d5 was part of that) for the default
template and in selfthinker/dokuwiki_template_starter@cc172e9f for the
starter template.
The new 'dokuwiki' template was based on the starter template but before
this change, so this never made it into the new default template.