Fix page's background colour in `export_html` mode
When using the `export_html` mode (page without surrounding layout)
the page's background colour is grey with a gradient.
That is the same background that the page with layout has
but different to the background the text is on.
What you would want is the same colour the text is on, which is white.