allow dialogs overflow in y-direction
In the dokuwiki template, an overflow:hidden was meant to keep oversized
content within the article area. However this can interfere with
pseudo-dropdowns on very small pages.

This changes the behavior to clip on x-axis only while keeping overflows
visible in the y direction.

This fixes a rare issue and should not have any effects on normal use
anyway.